Move relevant folders into lib garbo
