+++ src/libgarbo/index.ts
@@ -1,0 +1,7 @@
+import { makeValue } from "./value";
+import type { ValueFunctions } from "./value";
+import { WandererManager } from "./wanderer";
+import type { DraggableFight, WanderOptions } from "./wanderer";
+
+export { makeValue, WandererManager };
+export type { ValueFunctions, WanderOptions, DraggableFight };
