+++ src/libgarbo/wanderer/eightbit.ts
