+++ src/libgarbo/wanderer/freefight.ts
